backend/requirements.txt
langchain-openai
langchain-milvus
langchain-community
python-docx
numpy
pandas
pymilvus>=2.4.0
pymupdf
openai
unstructured
tiktoken
mysql-connector-python
pymysql
langchain-experimental
uvicorn
uvicorn[standard]
fastapi
genson
highcharts-core
sse-starlette
psycopg2-binary
alembic